requirements.txt
streamlit==1.22.0
matplotlib
numpy==1.23.5
pandas
scikit-learn
tensorflow==2.12.0
Pillow